requirements.txt
librosa==0.9.2
matplotlib<=3.7.0
numpy<2.0
pillow
panadas
scikit-learn
notebook
python-dotenv
tqdm
pandas
paramiko
tensorflow<2.11
sqlalchemy
pymysql